+++ src/mogr.js
@@ -1,0 +1,3 @@
+module.exports = function() {
+
+}
